refactor: ts migration for self-exclusion-article
